Fix bugs and resolve pending proposals from dogfooding phase [skip ci]
Bugs fixed:
- Regenerate 181l.h5msm, 1l17.h5msm, t4_lysozyme_L99A.h5msm: chain_name/chain_id
  were stored as 'nan'/'<NA>' due to generation with an older code version
- Fix float32 precision noise in coordinates after set() with string units:
  to_molsysmt_Structures.py now casts all float arrays to float64 on load from
  h5msm (which intentionally stores float32 for compactness)

Proposals implemented:
- form name aliases: digest_to_form now resolves singular/variant spellings
  (e.g. 'molsysmt.Structure' -> 'molsysmt.Structures') via a tolerance alias
  dict; these are NOT official form names and are not documented as such
- info() output: _to_python() helper strips numpy scalars to native Python types
  in all DataFrame columns across all element levels
- pdb: prefix alias: is_form and _extract_pdb_id now accept 'pdb:XXXX' and
  'PDB:XXXX' (case-insensitive throughout); 'pdb:' is NOT official syntax,
  accepted silently as a user-friendliness fallback only. Extended wwPDB IDs
  (pdb_0000XXXX) already supported; case-insensitive matching added for all
  prefix variants including 'PDB_0000XXXX'. All string_pdb_id converters now
  use the shared _extract_pdb_id() helper instead of ad-hoc replace calls.

Tests: 27 new tests covering is_form, _extract_pdb_id, digest_to_form aliases,
and info() native Python type output.

+++ molsysmt/form/molsysmt_H5MSMFileHandler/to_molsysmt_Structures.py
@@ -22,25 +22,25 @@ def to_molsysmt_Structures(item, atom_indices='all', structure_indices='all', sk
 
     # Coordinates
     coordinates_unit = structures_ds['coordinates'].attrs.get('unit', 'nm')
-    
+
     if is_all(structure_indices):
         if is_all(atom_indices):
-            tmp_item.coordinates = puw.quantity(structures_ds['coordinates'][:], coordinates_unit)
+            tmp_item.coordinates = puw.quantity(structures_ds['coordinates'][:].astype(np.float64), coordinates_unit)
         else:
-            tmp_item.coordinates = puw.quantity(structures_ds['coordinates'][:, atom_indices, :], coordinates_unit)
+            tmp_item.coordinates = puw.quantity(structures_ds['coordinates'][:, atom_indices, :].astype(np.float64), coordinates_unit)
     else:
         if is_all(atom_indices):
-            tmp_item.coordinates = puw.quantity(structures_ds['coordinates'][structure_indices, :, :], coordinates_unit)
+            tmp_item.coordinates = puw.quantity(structures_ds['coordinates'][structure_indices, :, :].astype(np.float64), coordinates_unit)
         else:
-            tmp_item.coordinates = puw.quantity(structures_ds['coordinates'][np.ix_(structure_indices, atom_indices, [0,1,2])], coordinates_unit)
+            tmp_item.coordinates = puw.quantity(structures_ds['coordinates'][np.ix_(structure_indices, atom_indices, [0,1,2])].astype(np.float64), coordinates_unit)
 
     # Box
     if 'box' in structures_ds and structures_ds['box'].shape[0] > 0:
         box_unit = structures_ds['box'].attrs.get('unit', 'nm')
         if is_all(structure_indices):
-            tmp_item.box = puw.quantity(structures_ds['box'][:], box_unit)
+            tmp_item.box = puw.quantity(structures_ds['box'][:].astype(np.float64), box_unit)
         else:
-            tmp_item.box = puw.quantity(structures_ds['box'][structure_indices, :, :], box_unit)
+            tmp_item.box = puw.quantity(structures_ds['box'][structure_indices, :, :].astype(np.float64), box_unit)
     else:
         tmp_item.box = None
 
@@ -49,14 +49,14 @@ def to_molsysmt_Structures(item, atom_indices='all', structure_indices='all', sk
         b_factor_unit = structures_ds.attrs.get('b_factor_unit', 'nanometer**2')
         if is_all(structure_indices):
             if is_all(atom_indices):
-                tmp_item.b_factor = puw.quantity(structures_ds['b_factor'][:], b_factor_unit)
+                tmp_item.b_factor = puw.quantity(structures_ds['b_factor'][:].astype(np.float64), b_factor_unit)
             else:
-                tmp_item.b_factor = puw.quantity(structures_ds['b_factor'][:, atom_indices], b_factor_unit)
+                tmp_item.b_factor = puw.quantity(structures_ds['b_factor'][:, atom_indices].astype(np.float64), b_factor_unit)
         else:
             if is_all(atom_indices):
-                tmp_item.b_factor = puw.quantity(structures_ds['b_factor'][structure_indices, :], b_factor_unit)
+                tmp_item.b_factor = puw.quantity(structures_ds['b_factor'][structure_indices, :].astype(np.float64), b_factor_unit)
             else:
-                tmp_item.b_factor = puw.quantity(structures_ds['b_factor'][np.ix_(structure_indices, atom_indices)], b_factor_unit)
+                tmp_item.b_factor = puw.quantity(structures_ds['b_factor'][np.ix_(structure_indices, atom_indices)].astype(np.float64), b_factor_unit)
     else:
         tmp_item.b_factor = None
 
@@ -64,9 +64,9 @@ def to_molsysmt_Structures(item, atom_indices='all', structure_indices='all', sk
     if 'time' in structures_ds and structures_ds['time'].shape[0] > 0:
         time_unit = structures_ds['time'].attrs.get('unit', 'ps')
         if is_all(structure_indices):
-            tmp_item.time = puw.quantity(structures_ds['time'][:], time_unit)
+            tmp_item.time = puw.quantity(structures_ds['time'][:].astype(np.float64), time_unit)
         else:
-            tmp_item.time = puw.quantity(structures_ds['time'][structure_indices], time_unit)
+            tmp_item.time = puw.quantity(structures_ds['time'][structure_indices].astype(np.float64), time_unit)
     else:
         tmp_item.time = None
